backend/requirements.txt
Django==5.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
django-cors-headers==4.4.0
django-environ==0.11.2
psycopg2-binary==2.9.9
gunicorn==22.0.0
reportlab==4.1.0
